requirements.txt
Django==1.9.4
django-filter==0.12.0
django-jet==0.1.4
djangorestframework==3.3.2
Markdown==2.6.5
MySQL-python==1.2.5
pep257==0.7.0
pep8==1.7.0
virtualenv==15.0.0
